test(conformance): split BleConformanceTest into focused files
538-line monolith split into 6 focused abstract classes:
- ScanConformanceTest (scan)
- ConnectionConformanceTest (connect/disconnect/reconnect/service)
- GattConformanceTest (discovery/read/write/notification)
- L2capConformanceTest (channel & listener)
- PhyConformanceTest (PHY selection & updates)

BleConformanceTest retains only factory methods (buildScanner/buildPeripheral).
JvmBleConformanceTest replaced by 5 platform-specific runners.
All 22 conformance tests pass.
