+++ src/jvmTest/kotlin/com/atruedev/kmpble/conformance/JvmL2capConformanceTest.kt
@@ -1,0 +1,10 @@
+package com.atruedev.kmpble.conformance
+
+/**
+ * JVM-specific L2CAP conformance test runner.
+ *
+ * Inherits all [L2capConformanceTest] test cases and runs them
+ * with the default [FakeScanner]/[FakePeripheralBuilder] factories.
+ * Run: ./gradlew :jvmTest --tests "*JvmL2capConformance*"
+ */
+public class JvmL2capConformanceTest : L2capConformanceTest()
